+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [1.11.0] 2025-10-14
+
+### Added
+
+- Write-LogEvent: Writes event log entry
+
+### Changed
+
+- Start-PSv2: uses Write-LogEvent to write to Event Log
+
 ## [1.10.0] 2025-08-26
 
 ### Added